+++ tumeplayMobile/src/components/Journey/Journey_modules/BottomAction.js
@@ -26,6 +26,7 @@ const BottomAction = ({style, selectedModule}) => {
         clearModuleData: true,
         retry: done,
         from_journey: true,
+        firstTry: true,
       });
     }
   };